requirements.txt
httpx
python-dotenv
typer[all]
pydantic 